+++ src/types/index.ts
@@ -403,6 +403,7 @@ export interface TreeGraphData {
     [key: string]: ShapeStyle;
   };
   stateStyles?: StateStyles;
+  [key: string]: unknown;
 }
 
 export interface NodeConfig extends ModelConfig {